+++ lib/suspenders/generators/app_generator.rb
@@ -34,7 +34,6 @@ module Suspenders
       invoke :setup_development_environment
       invoke :setup_test_environment
       invoke :setup_production_environment
-      invoke :setup_staging_environment
       invoke :setup_secret_token
       invoke :create_suspenders_views
       invoke :configure_app
@@ -116,11 +115,6 @@ module Suspenders
       build :enable_rack_canonical_host
       build :enable_rack_deflater
       build :setup_asset_host
-    end
-
-    def setup_staging_environment
-      say 'Setting up the staging environment'
-      build :setup_staging_environment
     end
 
     def setup_secret_token