+++ templates/smtp.rb
@@ -7,3 +7,7 @@ SMTP_SETTINGS = {
   port: "587",
   user_name: ENV.fetch("SMTP_USERNAME")
 }
+
+if ENV["EMAIL_RECIPIENTS"].present?
+  Mail.register_interceptor RecipientInterceptor.new(ENV["EMAIL_RECIPIENTS"])
+end