+++ templates/staging.rb
@@ -1,5 +1,0 @@
-require_relative "production"
-
-Mail.register_interceptor(
-  RecipientInterceptor.new(ENV.fetch("EMAIL_RECIPIENTS"))
-)